test: route daemon channel surfaces through channel shards
