+++ crates/daemon/tests/integration/mod.rs
@@ -7,8 +7,10 @@ mod architecture;
 mod ask_and_spec_cli_root;
 mod ask_cli;
 mod audit_cli_root;
-mod channel_surfaces_json;
-mod channel_surfaces_text;
+mod channel_catalog_json;
+mod channel_catalog_text;
+mod channel_plugin_bridge_json;
+mod channel_plugin_bridge_text;
 mod chat_cli;
 mod cli_tests;
 mod doctor_feishu;
